Add hide counter setting; stat calc; moved counter init into stat display
Stat calc works off a "first instance" of initial stats (cached with the counter). Clicking reset will reset the initial stats to the current instantiated version. Clicking on any of the stats will perform the calculation and show the max underneath (as opposed to auto calculation, which actually slowed down plugin loading). Addressed possible race condition with where counter was being instantiated versus when it was expecting it to be used.